[MOV] hr_skills: Split model files according to coding guidelines
