+++ addons/hr_skills/models/hr_employee.py
@@ -1,0 +1,29 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
+from odoo import api, fields, models
+
+
+class Employee(models.Model):
+    _inherit = 'hr.employee'
+
+    resume_line_ids = fields.One2many('hr.resume.line', 'employee_id', string="Resumé lines")
+    employee_skill_ids = fields.One2many('hr.employee.skill', 'employee_id', string="Skills")
+
+    @api.model_create_multi
+    def create(self, vals_list):
+        res = super(Employee, self).create(vals_list)
+        if self.env.context.get('salary_simulation'):
+            return res
+        resume_lines_values = []
+        for employee in res:
+            line_type = self.env.ref('hr_skills.resume_type_experience', raise_if_not_found=False)
+            resume_lines_values.append({
+                'employee_id': employee.id,
+                'name': employee.company_id.name or '',
+                'date_start': employee.create_date.date(),
+                'description': employee.job_title or '',
+                'line_type_id': line_type and line_type.id,
+            })
+        self.env['hr.resume.line'].create(resume_lines_values)
+        return res
