+++ addons/hr_skills/models/hr_employee_public.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
+from odoo import fields, models
+
+
+class EmployeePublic(models.Model):
+    _inherit = 'hr.employee.public'
+
+    resume_line_ids = fields.One2many('hr.resume.line', 'employee_id', string="Resumé lines")
+    employee_skill_ids = fields.One2many('hr.employee.skill', 'employee_id', string="Skills")
