+++ addons/hr_skills/models/hr_resume_line_type.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
+from odoo import fields, models
+
+
+class ResumeLineType(models.Model):
+    _name = 'hr.resume.line.type'
+    _description = "Type of a resumé line"
+    _order = "sequence"
+
+    name = fields.Char(required=True)
+    sequence = fields.Integer('Sequence', default=10)
